requirements.txt
Flask
pandas
numpy
matplotlib
scikit-learn==1.0.2
datetime
mysql-connector-python>=8.0.0

